@@ -1,13 +1,17 @@
 'use strict'
 
+process.env.NODE_ENV = 'development'
+
 const chalk = require('chalk')
 const electron = require('electron')
 const path = require('path')
 const { say } = require('cfonts')
 const { spawn } = require('child_process')
+const config = require('../_config')
 const webpack = require('webpack')
 const WebpackDevServer = require('webpack-dev-server')
 const webpackHotMiddleware = require('webpack-hot-middleware')
+const Portfinder = require("portfinder")
 
 const mainConfig = require('./webpack.main.config')
 const rendererConfig = require('./webpack.renderer.config')
@@ -19,7 +23,7 @@ let hotMiddleware
 function logStats (proc, data) {
   let log = ''
 
-  log += chalk.yellow.bold(`┏ ${proc} Process ${new Array((19 - proc.length) + 1).join('-')}`)
+  log += chalk.yellow.bold(`┏ ${proc} ${config.dev.chineseLog ? '编译过程' : 'Process'} ${new Array((19 - proc.length) + 1).join('-')}`)
   log += '\n\n'
 
   if (typeof data === 'object') {
@@ -34,58 +38,87 @@ function logStats (proc, data) {
   }
 
   log += '\n' + chalk.yellow.bold(`┗ ${new Array(28 + 1).join('-')}`) + '\n'
-
   console.log(log)
+}
+
+function removeJunk (chunk) {
+  if (config.dev.removeElectronJunk) {
+    // Example: 2018-08-10 22:48:42.866 Electron[90311:4883863] *** WARNING: Textured window <AtomNSWindow: 0x7fb75f68a770>
+    if (/\d+-\d+-\d+ \d+:\d+:\d+\.\d+ Electron(?: Helper)?\[\d+:\d+] /.test(chunk)) {
+      return false;
+    }
+
+    // Example: [90789:0810/225804.894349:ERROR:CONSOLE(105)] "Uncaught (in promise) Error: Could not instantiate: ProductRegistryImpl.Registry", source: chrome-devtools://devtools/bundled/inspector.js (105)
+    if (/\[\d+:\d+\/|\d+\.\d+:ERROR:CONSOLE\(\d+\)\]/.test(chunk)) {
+      return false;
+    }
+
+    // Example: ALSA lib confmisc.c:767:(parse_card) cannot find card '0'
+    if (/ALSA lib [a-z]+\.c:\d+:\([a-z_]+\)/.test(chunk)) {
+      return false;
+    }
+  }
+
+
+  return chunk;
 }
 
 function startRenderer () {
   return new Promise((resolve, reject) => {
-    rendererConfig.entry.renderer = [path.join(__dirname, 'dev-client')].concat(rendererConfig.entry.renderer)
+    rendererConfig.mode = 'development'
+    Portfinder.basePort = config.dev.port || 9080
+    Portfinder.getPort((err, port) => {
+      if (err) {
+        reject("PortError:" + err)
+      } else {
+        const compiler = webpack(rendererConfig)
+        hotMiddleware = webpackHotMiddleware(compiler, {
+          log: false,
+          heartbeat: 2500
+        })
 
-    const compiler = webpack(rendererConfig)
-    hotMiddleware = webpackHotMiddleware(compiler, { 
-      log: false, 
-      heartbeat: 2500 
-    })
-
-    compiler.plugin('compilation', compilation => {
-      compilation.plugin('html-webpack-plugin-after-emit', (data, cb) => {
-        hotMiddleware.publish({ action: 'reload' })
-        cb()
-      })
-    })
-
-    compiler.plugin('done', stats => {
-      logStats('Renderer', stats)
-    })
-
-    const server = new WebpackDevServer(
-      compiler,
-      {
-        contentBase: path.join(__dirname, '../'),
-        quiet: true,
-        setup (app, ctx) {
-          app.use(hotMiddleware)
-          ctx.middleware.waitUntilValid(() => {
-            resolve()
+        compiler.hooks.afterEmit.tap('afterEmit', () => {
+          hotMiddleware.publish({
+            action: 'reload'
           })
-        }
-      }
-    )
+        })
 
-    server.listen(9080)
+        compiler.hooks.done.tap('done', stats => {
+          logStats('Renderer', stats)
+        })
+
+        const server = new WebpackDevServer(
+          compiler, {
+          contentBase: path.join(__dirname, '../'),
+          quiet: true,
+          before (app, ctx) {
+            app.use(hotMiddleware)
+            ctx.middleware.waitUntilValid(() => {
+              resolve()
+            })
+          }
+        }
+        )
+
+        process.env.PORT = port
+        server.listen(port)
+
+      }
+    })
+
   })
 }
 
 function startMain () {
-  return new Promise((resolve, reject) => {
-    mainConfig.entry.main = [path.join(__dirname, '../src/main/index.dev.js')].concat(mainConfig.entry.main)
-
+  return new Promise((resolve) => {
+    mainConfig.mode = 'development'
     const compiler = webpack(mainConfig)
 
-    compiler.plugin('watch-run', (compilation, done) => {
-      logStats('Main', chalk.white.bold('compiling...'))
-      hotMiddleware.publish({ action: 'compiling' })
+    compiler.hooks.watchRun.tapAsync('watch-run', (compilation, done) => {
+      logStats(`${config.dev.chineseLog ? '主进程' : 'Main'}`, chalk.white.bold(`${config.dev.chineseLog ? '正在处理资源文件...' : 'compiling...'}`))
+      hotMiddleware.publish({
+        action: 'compiling'
+      })
       done()
     })
 
@@ -95,7 +128,7 @@ function startMain () {
         return
       }
 
-      logStats('Main', stats)
+      logStats(`${config.dev.chineseLog ? '主进程' : 'Main'}`, stats)
 
       if (electronProcess && electronProcess.kill) {
         manualRestart = true
@@ -114,13 +147,25 @@ function startMain () {
 }
 
 function startElectron () {
-  electronProcess = spawn(electron, ['--inspect=5858', path.join(__dirname, '../dist/electron/main.js')])
+  var args = [
+    '--inspect=5858',
+    path.join(__dirname, '../dist/electron/main.js')
+  ]
+
+  // detect yarn or npm and process commandline args accordingly
+  if (process.env.npm_execpath.endsWith('yarn.js')) {
+    args = args.concat(process.argv.slice(3))
+  } else if (process.env.npm_execpath.endsWith('npm-cli.js')) {
+    args = args.concat(process.argv.slice(2))
+  }
+
+  electronProcess = spawn(electron, args)
 
   electronProcess.stdout.on('data', data => {
-    electronLog(data, 'blue')
+    electronLog(removeJunk(data), 'blue')
   })
   electronProcess.stderr.on('data', data => {
-    electronLog(data, 'red')
+    electronLog(removeJunk(data), 'red')
   })
 
   electronProcess.on('close', () => {
@@ -129,20 +174,23 @@ function startElectron () {
 }
 
 function electronLog (data, color) {
-  let log = ''
-  data = data.toString().split(/\r?\n/)
-  data.forEach(line => {
-    log += `  ${line}\n`
-  })
-  if (/[0-9A-z]+/.test(log)) {
-    console.log(
-      chalk[color].bold('┏ Electron -------------------') +
-      '\n\n' +
-      log +
-      chalk[color].bold('┗ ----------------------------') +
-      '\n'
-    )
+  if (data) {
+    let log = ''
+    data = data.toString().split(/\r?\n/)
+    data.forEach(line => {
+      log += `  ${line}\n`
+    })
+    if (/[0-9A-z]+/.test(log)) {
+      console.log(
+        chalk[color].bold(`┏ ${config.dev.chineseLog ? '主程序日志' : 'Electron'} -------------------`) +
+        '\n\n' +
+        log +
+        chalk[color].bold('┗ ----------------------------') +
+        '\n'
+      )
+    }
   }
+
 }
 
 function greeting () {
@@ -160,19 +208,20 @@ function greeting () {
       space: false
     })
   } else console.log(chalk.yellow.bold('\n  electron-vue'))
-  console.log(chalk.blue('  getting ready...') + '\n')
+  console.log(chalk.blue(`${config.dev.chineseLog ? '  准备启动...' : '  getting ready...'}`) + '\n')
 }
 
-function init () {
+async function init () {
   greeting()
 
-  Promise.all([startRenderer(), startMain()])
-    .then(() => {
-      startElectron()
-    })
-    .catch(err => {
-      console.error(err)
-    })
+  try {
+    await startRenderer()
+    await startMain()
+    await startElectron()
+  } catch (error) {
+    console.error(error)
+  }
+
 }
 
 init()
