@@ -1,15 +1,26 @@
 <template>
-  <div>
-      <router-view></router-view>
+  <div id="app">
+    <router-view></router-view>
   </div>
 </template>
 
 <script>
 export default {
-
-}
+};
 </script>
 
 <style>
-
+* {
+  user-select: none !important;
+}
+html {
+  background-color: #333 !important;
+}
+body {
+  background-color: #333 !important;
+  overflow: hidden !important;
+}
+.ant-modal {
+  width: 300px !important;
+}
 </style>
