@@ -6,16 +6,16 @@ import router from './router'
 import Antd from 'ant-design-vue'
 import 'ant-design-vue/dist/antd.css'
 
-import '@/assets/css/iconfont.css'
+// 日志
+import './error'
+
 Vue.use(Antd)
 
-if (!process.env.IS_WEB) Vue.use(require('vue-electron'))
-
 Vue.config.productionTip = false
-
 /* eslint-disable no-new */
 new Vue({
   el: '#app',
   router,
   ...App
 })
+
